requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
aiosqlite==0.17.0
asgiref==3.5.2
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cycler==0.11.0
Django==4.1
flatbuffers==2.0.7
fonttools==4.37.1
frozenlist==1.3.1
gast==0.4.0
google-auth==2.11.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.49.1
h5py==3.7.0
idna==3.3
image==1.5.33
joblib==1.2.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
multidict==6.0.2
neuralintents==0.0.4
nextcord==2.2.0
nltk==3.7
numpy==1.23.2
oauthlib==3.2.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
protobuf==3.19.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
tqdm==4.64.1
typing_extensions==4.3.0
tzdata==2022.2
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
yarl==1.8.1